requirements.txt
aiohttp
pillow
requests